add stSetPT keyword.  syntax> stSetPT:att:def:altCondition:altAtt:altDef:Description
This should consolidate code using StaticEffect keyword and GameActionUtil/StaticEffects custom code.

Also, the card.txt now accepts the * or 2+* for displaying in the DeckEditor.  The game will still display the actual numbers.

The one todo is that card overlays are not updated correctly when the card is in hand.  The actual stats are updated correctly however.

converted Nightmare and Angry Mob as examples.

git-svn-id: http://svn.slightlymagic.net/forge/trunk/res/cardsfolder@4201 269b9781-a132-4a9b-9d4e-f004f1b56b58