add the basics of a system to handle cards with Imprint.  You need to use AF abilities and set Imprint$ True in the AF line in the card.  If a card can imprint more than once and needs to forget previously imprinted cards, set Unimprint$ True.  Imprinted cards can be accessed like: *SVar:X:Imprinted$CardManaCost *Defined$ Imprinted
Also, check the card detail panel to see which cards a host card is imprinting.

Caution: if the AF does not use Target (like ChangeZone with Hand origin), it may not work, and code will need to be added to support those.  Please post in the Imprint thread in the dev forum if it isn't working for you.

git-svn-id: http://svn.slightlymagic.net/forge/trunk@8624 269b9781-a132-4a9b-9d4e-f004f1b56b58